requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altgraph==0.17.4
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
docopt==0.6.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pytest==8.3.2
pytest-watch==4.2.0
python-dotenv==1.0.1
rich==13.7.1
setuptools==72.2.0
sniffio==1.3.1
textual==0.76.0
textual-dev==1.5.1
typing_extensions==4.12.2
uc-micro-py==1.0.3
watchdog==4.0.2
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
